requirements.txt
einops==0.8.0
h5py==3.12.1
julius==0.2.7
librosa==0.10.2.post1
lightning==2.5.0.post0
matplotlib==3.9.0
natsort==8.4.0
nnAudio==0.3.3
numpy==1.*
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.19.2
onnxscript==0.1.0.dev20241029
onnxsim==0.4.36
praat-parselmouth==0.4.5
pydub==0.25.1
pyreaper==0.0.10
pyworld==0.3.4
soundfile==0.12.1
tensorboard
torch
torchaudio
torchmetrics[audio]==1.5.1
tqdm==4.66.6
